note, late fine and overtiem api

+++ migrations/20201016131243-create-attendance.js
@@ -35,7 +35,7 @@ module.exports = {
         type: Sequelize.TIME
       },
       overtime_pay: {
-        type: Sequelize.DOUBLE
+        type: Sequelize.DECIMAL
       },
       late_fine_hours: {
         type: Sequelize.TIME
